decoder_raw: Fix generation of WHERE clause
The attribute value used to determine what is the value needed to identify
a row when a relation uses an index as replica identity was incorrect.

Report from bwtakacy, who also submitted a patch that I actually did not
use as it seems better to pass the real attribute number of relation
to print_where_clause_item. Regression tests are from me.