pg_syncrep_state: First implementation
This module can be used to monitor precisely the state of each backend
regarding synchronous replication without needing to have a look at ps
display for this purpose, pg_stat_activity being not enough verbose in
this case.

This module, having one new function and a view wrapping it, is recommended
to be used joined with pg_stat_activity to monitor the amount of lock time
a qucery has been spent on. This does not necessarily correspond to the
beginning point when synchronous replication lock has been taken but some
application can satisfy themselves by taking as a base the time commit
query has begun before becoming stuck.