pg_synrep_state: Improve detection of inactive backends
The process array list does not reset backends already stopped hence add
a check depending on kill(0) that a given process already exists before
returning its value to caller. It may be possible that the PID saved for
the saved but dead entry in the list is taken by another process though.