pg_trunc2del: Initial commit
pg_trunc2del aims at replacing on-the-fly TRUNCATE statements by an
equivalent DELETE and executing it transparently to the user. This
extension has a couple of limits relation to triggers and other
TRUNCATE options like CASCADE but this seems enough for the purpose
of demonstration.

Original idea of this extension from Julien Rouhaud, though I didn't
reuse his idea and relied on SPI to provide a transparent query run
from the user prospective.