pg_panic: Initial commit
pg_panic is an extension triggering PANIC, or a server-wide failure, or
the worst possible level possible in a Postgres instance in a non
deterministic way using a user-settable luck factor. This extension may be
useful to emulate FS or OS-level failures that could happen when Postgres
is running and check the robustness of an upper-level application layer.

Original idea from Julien Rouhaud and Michael Paquier, with extra ideas
from Nicolas Thauvin.