jsonlog: Remove junk prefixes in stderr output
Those caused the json entries to be generated with an annoying junk prefix
that was port of syslogger format, and making them unusable if taken as
such when they were redirected to stdout. Note that when the logging
collector is enabled, the first log entries generated by the postmaster
which are unconditionally logged into stderr will still have some junk,
there is not much doable here as there is a control flag, but localized
only in the postmaster.

Per report from Guillaume Polaert, though I was aware of that for months,
and lacked motivation to address it.