pg_swap_pages: Add template extension using WAL generic interface
This is a 9.6-only module, which makes use of the generic WAL interface
allowing one to log pages of a relation at will. This modules corrupts
a system because it blindly switches the contents of two pages, and on
top of that does the same for nodes replaying those changes.