decoder_raw: Handle Infinity and NaN in numeric and float
When Infinity, -Infinity or NaN shows up in a float or numeric value,
they need to be quoted. Infinity and -Infinity cannot appear for
Numeric, only float.

At the same time add regression tests for that, and boolean values
that got a special handling as well.

Patch by Magnus Hagander.