pg_sasl_prepare: reduce size of conversion table
Characters with a class of 0 and no decompositions don't really need
to be in the table and any abscence of such characters in the table
would mean that they have the following limitations. This reduces the
size of the table from 2.4MB to 570kB, which is already a good win.