pg_sasl_prepare: Move decomposition tables into sub-tables
The main table tracks the length of each decomposition, which is then
used with each character code to look at the decomposition set on a
set of tables divided by decomposition size. This gets the binary size
from 2.4MB to 120kB, or 20 times less. I am not sure if I could get
that further down, no space is wasted now.