pg_wal_blocks: Fix compilation with v11
The WAL segment size is something that can be customized at initialization
time in v11, and was previously decided at compilation time. This makes the
module able to work with a default segment size of 16MB. In order to
support more sizes, the segment size needs to be passed as an argument.