jsonlog: Add the statement to the log
This is printed in the regular logs as "STATEMENT" and in CSV logs the
column is (confusingly) documented as "query". Follow the regular logs
convention which is more accurate and call it "statement".

In passing change the appendJSONLiteral to take const char* to avoid
warnings about casting away constness of debug_query_string..