jsonlog: Fix handling of log_min_messages for "LOG"-level messages
Postgres's handling of LOG level messages relies on is_log_level_output
in elog.c, which is here to determine if a message should be logged or
not.  LOG messages are treated as between DEBUG1 and INFO for the client
however for the server log they're treated as equal to ERROR.

In order to get a consistent handling as with the backend, copy the
static is_log_level_output function from elog.c to implement the same
behavior.

One additional change is COMMERROR is used instead of LOG_SERVER_ONLY
which is an alias of the former introduced in 9.6.  This is to make this
copied routine compatible with older server versions.

Patch and report by Greg Stark.