jsonlog: Fix handling of timestamps in entries
This fixes two issues:
1) The formatted log value would remain the same for the life duration
of a single backend, so regenerate the timestamp for each entry as
appropriate.  It could be possible to reinitialize the static variable
in charge of saving the formatted timestamp, but as this code path is
taken only once per elog entry that is not worth the complication.
2) Fix corruption in dates generated because of an incorrect number of
spaces in timestamp parsing.

While at it, make setup_formatted_log_time more consistent with
upstream.  Upstream commit 586dd5d6 has replaced strncpy calls with
safer memcpy calls, so do the same.  The variables used are renamed to
match upstream as well.

Report of the two bugs with Greg Stark, who also sent a patch to get
both issues addressed.  The changes to match upstream code are from me.