jsonlog: Add statement and cursor position
Statements are printed in the regular logs as "STATEMENT" and in CSV
logs the column is documented as "query", which is confusing. So follow
the regular logs convention which is more accurate and call it
"statement".

The cursor position is added as well, and in passing change the
appendJSONLiteral to take const char* to avoid warnings about casting
away const-ness of debug_query_string.

Report and patch by Greg Stark.