jsonlog: Ignore log_timezone and use the "Z" time zone format suffix
This is the format the popular json filtering tool, jq, can handle and
it is the format Javascript produces and expects like for Date's toJSON.
As the format is decided by ISO 8601 which is rather spread, that may
prove to be more consistent at the end.

If there are complains about the timestamp format in the future, we
could always consider that optional with say a switch in charge of
switching between for example the "classic" format used before this
commit and the Javascript format which is introduced here. To keep the
code simple just use the latter for now.

Note also that this enforces UTC.

Patch and suggestions by Greg Stark.