pg_checksums: Ignore temporary files when scanning entries
Those temporary files do not include data checksummed, causing the tool
to fail if it tries to work on them, be it to verify or enable
checksums.  Those files may not be immediately removed by Postgres at
the beginning of recovery even if the instance has been forcibly
killed.

Add some tests at the same time covering this case.  While on it, adjust
header order to be more consistent with PostgreSQL policy on the
matter.