jsonlog: Fix generation of JSON blobs sent to stderr
An incorrect data length was passed down to the write() call sending the
JSON blobs to stderr, which would happen for example when
logging_collector is disabled.  This resulted in some junk data
generated.

Testing this module without the logging_collector and the output sent to
stderr proves that this fixes the problem.

Report and patch by Gilles Darold.