jsonlog: Fix compilation with Postgres 15
Upstream has begun to publish in ac7c807 a couple of routines whose name
match with what this module provides.  It could be possible to remove
those routines all together, but as we need to care about the
compilation with older upstream versions, just rename the existing
routines to avoid any conflicts, prefixing them with "jsonlog_".