signal_rmgr: Custom WAL resource manager to send signal to standbys
This module can be loaded via shared_preload_libraries to register a
custom WAL resource manager, then the SQL function signal_rmgr() can be
used to generate signals that will be executed on the standbys replaying
these records:
- SIGHUP
- SIGKILL
- SIGTERM
- SIGINT

The signals are sent to a node's postmaster.  A TAP test is added, while
on it.