custom_wal: First implementation
This is a module that works as a thin wrapper of LogLogicalMessage().
PostgreSQL core already provides an equivalent of this function, but
this is able to generate a series of WAL records in a single function
call, which is very useful to stress the concurrency of WAL records
generated particularly the LWLocks for WAL insertion.