Plug in WaitEventExtensionNew() to all the modules
This new routine, introduced by upstream commit af720b4, allows a module
to define custom strings for extension wait events.  Reporting a wait
event for a bgworker requires a database connection, so this does not
work for some of the modules changed here, still it shows the pattern to
follow to be able to achieve the change.