blackhole_am: Add NOTICE messages to callbacks
This change makes it easier to play with the extension, which is the
reason why it exists to get a feeling of the table AM interface.  It is
also trivial to change the NOTICE to other log levels, if needed.

The regression tests are updated to show the NOTICE messages now
generated.

Per pull request #46.

Author: Onder Kalaci