jsonlog: Add jsonlog.service
If this GUC is defined, all the log entries gain a new key/value pair
with "service" as key, and the GUC defined as value.

If multiple services, PostgreSQL being one, emit logs in a unified JSON
format to stderr/stdout and pipe this information to a single, central,
location, the "service" key/value is useful to know from which service a
given JSON entry is from.

Author: Julien Rouhaud

Per pull request #51.