Add a new jsonlog.service GUC
If the GUC is defined, each logs will be enitted with a new key names "service"
values with the configured value.

This can be useful in cloud environment where the best practice is to emit
everything on stdout or stderr.  If multiple processes can emit logs a service
key is needed to identify the service associated to each log.