jsonlog: Make the service key identifier configurable
Some tools like datadog have a hardcoded "service" identifier for all the
ingested logs, which cannot be modified.  Some other tools might have similar
incompatibility with other values, so making the service key user configurable
should make sure that jsonlog can be used in any environment.

This commit adds a new "jsonlog.service_key", which default to "component" as
"service" is known to be problematic with a commonly used tool, and renames the
previous "jsonlog.servce" to "jsonlog.service_value".