hint_parser: Silence warnings with -Wmissing-variable-declarations
Upstream PostgreSQL has added the -Wmissing-variable-declarations
compilation switch by default since 66188912566b.  This is a workaround
to avoid these warnings.

A better long-term approach would be to switch to pure parsers in bison,
so as global variables are not required, but this may not be worth the
move for this template module.